refactor: 게임 재시작 여부의 validation 코드를 이동하다.
