+++ src/main/java/baseball/GameValidation.java
@@ -1,0 +1,12 @@
+package baseball;
+
+import baseball.resource.AskRestartGame;
+
+public final class GameValidation {
+    private String userInput;
+
+    public void setUserInput(String userInput) {
+        this.userInput = userInput;
+    }
+
+}
